I feel like I'm doing things the hard way.
Namely, Atlas's contain textures, which are attached in the editor to the
main object and then passed into the world.  In AOQT1 this is necessary
because things get loaded at the start and then just hung onto in the
main object.  So we need the texture in the atlas defined so that game objects
know what to draw.  There's some additional code just in case atlases are
different, but in reality they're all the same size and have the same number
of sprites.

In Unity though, you already tell objects in the editor what texture they
need to be using.  So while they need to know the (fixed) size and sprite
count of their atlas, they already have a reference to what the texture
actually is.  Because they're already doing the drawStuff(texture, location)
code for me under the hood.

So it seems like, since all atlases will be the same, the obvious choice is
to have the Atlas object just provide static methods, and what atlas an object
has will be determined by just what texture/material is attached.  This does
make it slightly weirder to use multiple different textures, but... meh.